[#64] refactor: API Campaign Reporrt 를 Campaign API 로 통합
