+++ live-commerce-api/src/main/java/org/giglab/live/commerce/api/dto/campaign/CreateCampaignReportRequest.java
@@ -1,10 +1,10 @@
-package org.giglab.live.commerce.api.dto.broadcast;
+package org.giglab.live.commerce.api.dto.campaign;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 
-public record SaveCampaignReportRequest(
+public record CreateCampaignReportRequest(
     @NotBlank String chatRoomId,
     int totalViewers,
     int peakConcurrent,
