--- live-commerce-api/src/main/java/org/giglab/live/commerce/api/facade/CampaignReportFacade.java
@@ -1,19 +1,0 @@
-package org.giglab.live.commerce.api.facade;
-
-import lombok.RequiredArgsConstructor;
-import org.giglab.live.commerce.api.dto.broadcast.SaveCampaignReportRequest;
-import org.giglab.live.commerce.api.mapper.campaign.CampaignReportMapper;
-import org.giglab.live.commerce.core.campaign.application.CampaignReportService;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class CampaignReportFacade {
-
-  private final CampaignReportService campaignReportService;
-  private final CampaignReportMapper campaignReportMapper;
-
-  public void save(SaveCampaignReportRequest request) {
-    campaignReportService.save(campaignReportMapper.toSaveCampaignReportCommand(request));
-  }
-}
