+++ live-commerce-api/src/main/java/org/giglab/live/commerce/api/mapper/campaign/CampaignReportMapper.java
@@ -1,11 +1,11 @@
 package org.giglab.live.commerce.api.mapper.campaign;
 
-import org.giglab.live.commerce.api.dto.broadcast.SaveCampaignReportRequest;
-import org.giglab.live.commerce.core.campaign.application.dto.SaveCampaignReportCommand;
+import org.giglab.live.commerce.api.dto.campaign.CreateCampaignReportRequest;
+import org.giglab.live.commerce.core.campaign.application.dto.CreateCampaignReportCommand;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
 public interface CampaignReportMapper {
 
-  SaveCampaignReportCommand toSaveCampaignReportCommand(SaveCampaignReportRequest request);
+  CreateCampaignReportCommand toSaveCampaignReportCommand(CreateCampaignReportRequest request);
 }
